+++ openpyxl2/formatting/tests/test_formatting.py
@@ -124,7 +124,7 @@ def test_conditional_formatting_read(datadir):
     reference_file = 'conditional-formatting.xlsx'
     wb = load_workbook(reference_file)
     ws = wb.active
-    rules = ws.conditional_formatting.cf_rules
+    rules = ws.conditional_formatting
     assert len(rules) == 30
 
     # First test the conditional formatting rules read